#97 Support Firefox multiprocess (electrolysis, e10s)
